bll/producer/requirements.txt
pycryptodome==3.20.0
python_bcrypt==0.3.2
SQLAlchemy==2.0.30
websockets==12.0
